livekit-agent/requirements.txt
livekit>=1.0.0
livekit-agents>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-elevenlabs>=1.0.0
python-dotenv>=1.0.0
